Extract logic to the independent classes
